test: merge `Profiler` with `TimeUtil`
